requirements.txt
Django==4.0.3
psycopg2-binary==2.9.3
requests==2.27.1
djangorestframework==3.13.1
django-filter==22.1